+++ hadoop-ozone/common/src/main/java/org/apache/hadoop/ozone/audit/OMAction.java
@@ -88,7 +88,9 @@ public enum OMAction implements AuditAction {
 
   TENANT_ASSIGN_ADMIN,
   TENANT_REVOKE_ADMIN,
-  TENANT_LIST_USER;
+  TENANT_LIST_USER,
+
+  CREATE_SNAPSHOT;
 
   @Override
   public String getAction() {
