+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/SnapshotManager.java
@@ -1,86 +1,20 @@
 package org.apache.hadoop.ozone.om;
 
-import org.apache.commons.lang3.tuple.Pair;
-import org.apache.hadoop.hdds.conf.OzoneConfiguration;
-import org.apache.hadoop.hdds.scm.protocol.ScmBlockLocationProtocol;
-import org.apache.hadoop.hdds.scm.protocol.StorageContainerLocationProtocol;
 import org.apache.hadoop.hdds.utils.db.DBCheckpoint;
 import org.apache.hadoop.hdds.utils.db.RDBStore;
-import org.apache.hadoop.ipc.ProtobufRpcEngine;
-import org.apache.hadoop.ipc.Server;
-import org.apache.hadoop.ozone.OzoneConsts;
-import org.apache.hadoop.ozone.audit.AuditLogger;
-import org.apache.hadoop.ozone.audit.AuditLoggerType;
-import org.apache.hadoop.ozone.om.exceptions.OMException;
-import org.apache.hadoop.ozone.om.helpers.OmBucketInfo;
-import org.apache.hadoop.ozone.om.helpers.OmKeyArgs;
-import org.apache.hadoop.ozone.om.helpers.OmKeyInfo;
-import org.apache.hadoop.ozone.om.helpers.OmVolumeArgs;
-import org.apache.hadoop.ozone.om.request.OMClientRequest;
-import org.apache.hadoop.ozone.protocol.proto.OzoneManagerProtocolProtos;
-import org.apache.hadoop.ozone.security.acl.*;
-import org.apache.hadoop.security.UserGroupInformation;
-import org.apache.hadoop.util.ReflectionUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.net.InetAddress;
-import java.net.InetSocketAddress;
-import java.util.*;
-
-import static org.apache.hadoop.hdds.utils.HAUtils.getScmBlockClient;
-import static org.apache.hadoop.hdds.utils.HAUtils.getScmContainerClient;
-import static org.apache.hadoop.ozone.OzoneConfigKeys.*;
-import static org.apache.hadoop.ozone.OzoneConsts.OM_KEY_PREFIX;
-import static org.apache.hadoop.ozone.om.KeyManagerImpl.getRemoteUser;
-import static org.apache.hadoop.ozone.om.OMConfigKeys.OZONE_OM_VOLUME_LISTALL_ALLOWED;
-import static org.apache.hadoop.ozone.om.OMConfigKeys.OZONE_OM_VOLUME_LISTALL_ALLOWED_DEFAULT;
-import static org.apache.hadoop.ozone.om.OzoneManager.getACLAuthorizerInstance;
-import static org.apache.hadoop.ozone.om.OzoneManager.getS3Auth;
-import static org.apache.hadoop.ozone.om.exceptions.OMException.ResultCodes.DETECTED_LOOP_IN_BUCKET_LINKS;
-import static org.apache.hadoop.ozone.om.lock.OzoneManagerLock.Resource.BUCKET_LOCK;
-import static org.apache.hadoop.ozone.om.lock.OzoneManagerLock.Resource.VOLUME_LOCK;
 
 public class SnapshotManager {
+
   private static String getSnapshotDirName(String name, String mask) {
     return "-" + mask.replaceAll("/", "-") + "_" + name;
   }
+
   public static DBCheckpoint createSnapshot(OMMetadataManager omMetadataManager, String name, String mask)
       throws IOException {
     RDBStore store = (RDBStore) omMetadataManager.getStore();
     return store.getSnapshot(getSnapshotDirName(name, mask));
   }
 
-  public static class SnapshotMask {
-    public String getVolume() {
-      return volume;
-    }
-
-    public String getBucket() {
-      return bucket;
-    }
-
-    public String getPath() {
-      return path;
-    }
-
-    private String volume;
-    private String bucket;
-    private String path;
-    public SnapshotMask(String mask) {
-      String []names = mask.split(OM_KEY_PREFIX);
-      if (names.length > 0) {
-        volume = names[0];
-      }
-      if (names.length > 1) {
-        bucket = names[1];
-      }
-      
-      if (names.length > 2) {
-        path = String.join(OM_KEY_PREFIX,
-          Arrays.copyOfRange(names, 2, names.length));
-      }
-    }
-  }
 }
